✨ feat(analyzis error): Add custom errors when page is KO
