+++ api/application/middleware/analysis.py
@@ -1,8 +1,7 @@
+import api.domain.ecoindex.repository
 from ecoindex_scraper.models import WebPage
 from fastapi import HTTPException, Response, status
 from sqlmodel.ext.asyncio.session import AsyncSession
-
-import api.domain.ecoindex.repository
 
 
 async def validate_analysis_request(
