+++ api/application/status.py
@@ -1,0 +1,2 @@
+HTTP_520_ECOINDEX_TYPE_ERROR = 520
+HTTP_521_ECOINDEX_CONNECTION_ERROR = 521
