+++ api/main.py
@@ -1,23 +1,8 @@
-from db.engine import create_db_and_tables, is_database_online
-from fastapi import status
+from db.engine import create_db_and_tables
 from fastapi.applications import FastAPI
-from fastapi.middleware.cors import CORSMiddleware
-from fastapi.requests import Request
-from fastapi.responses import JSONResponse
-from fastapi_health import health
-from selenium.common.exceptions import WebDriverException
-from settings import (
-    CORS_ALLOWED_CREDENTIALS,
-    CORS_ALLOWED_HEADERS,
-    CORS_ALLOWED_METHODS,
-    CORS_ALLOWED_ORIGINS,
-)
 
 from api.domain.ecoindex import routes as ecoindex_routes
-from api.domain.health.chromedriver import is_chromedriver_healthy
 from api.domain.host import routes as host_routes
-from api.helper import format_exception_response
-from api.models.responses import ApiHealth
 
 app = FastAPI(
     title="Ecoindex API",
@@ -25,70 +10,17 @@ app = FastAPI(
     description="Ecoindex API enables you to perform ecoindex analysis of given web pages",
 )
 
-app.add_middleware(
-    CORSMiddleware,
-    allow_credentials=CORS_ALLOWED_CREDENTIALS,
-    allow_headers=CORS_ALLOWED_HEADERS,
-    allow_methods=CORS_ALLOWED_METHODS,
-    allow_origins=CORS_ALLOWED_ORIGINS,
-)
-
 app.include_router(router=ecoindex_routes.router)
 app.include_router(router=host_routes.router)
+
+from api.application.exception_handler import handle_exceptions
+from api.application.middleware.cors import add_cors_middleware
+from api.domain.health.route import add_healthcheck_route
 
 
 @app.on_event("startup")
 async def on_startup():
     await create_db_and_tables()
-
-
-@app.exception_handler(WebDriverException)
-async def handle_webdriver_exception(_: Request, exc: WebDriverException):
-    if "ERR_NAME_NOT_RESOLVED" in exc.msg:
-        return JSONResponse(
-            status_code=status.HTTP_502_BAD_GATEWAY,
-            content={
-                "detail": "This host is unreachable. Are you really sure of this url? 🤔"
-            },
-        )
-
-    if "ERR_CONNECTION_TIMED_OUT" in exc.msg:
-        return JSONResponse(
-            status_code=status.HTTP_504_GATEWAY_TIMEOUT,
-            content={
-                "detail": "Timeout reached when requesting this url. This is probably a temporary issue. 😥"
-            },
-        )
-
-    exception_response = await format_exception_response(exception=exc)
-    return JSONResponse(
-        content={"detail": exception_response.dict()},
-        status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-    )
-
-
-@app.exception_handler(RuntimeError)
-async def handle_screenshot_not_found_exception(_: Request, exc: FileNotFoundError):
-    return JSONResponse(
-        content={"detail": str(exc)},
-        status_code=status.HTTP_404_NOT_FOUND,
-    )
-
-
-@app.exception_handler(Exception)
-async def validation_exception_handler(_: Request, exc: Exception):
-    exception_response = await format_exception_response(exception=exc)
-    return JSONResponse(
-        content={"detail": exception_response.dict()},
-        status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-    )
-
-
-app.add_api_route(
-    path="/health",
-    endpoint=health([is_database_online, is_chromedriver_healthy]),
-    tags=["Infra"],
-    name="Get healthcheck",
-    description="Check health status of components of the API (database...)",
-    response_model=ApiHealth,
-)
+    await handle_exceptions()
+    await add_cors_middleware()
+    await add_healthcheck_route()
