+++ pyproject.toml
@@ -5,13 +5,13 @@ description = "REST API to expose Ecoindex"
 authors = ["Vincent Vatelot <vincent.vatelot@ik.me>"]
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.10"
 fastapi = "^0.85.2"
 python-environ = "^0.4.54"
 sqlmodel = "^0.0.7"
 fastapi-health = "^0.4.0"
 typer = "^0.6.1"
-ecoindex-scraper = "^2.4.1"
+ecoindex-scraper = "^2.6.0"
 uvicorn = {extras = ["standard"], version = "^0.19.0"}
 gunicorn = "^20.1.0"
 alembic = "^1.8.1"
@@ -19,7 +19,7 @@ alembic = "^1.8.1"
 [tool.poetry.dev-dependencies]
 black = "^22.3"
 aiosqlite = "^0.17.0"
-Faker = "^15.1.2"
+Faker = "^15.1.3"
 faker-enum = "^0.0.2"
 pytest-asyncio = "^0.20.1"
 pytest-mock = "^3.9.0"
